+++ src/Chaching/Drivers/PBiTerminal/Refund.php
@@ -195,7 +195,7 @@ class Refund extends \Chaching\Message
 		return $this->status;
 	}
 
-	private function request_server_url()
+	protected function request_server_url()
 	{
 		return ($this->environment === Chaching::SANDBOX)
 			? 'https://secureshop-test.firstdata.lv:8443/ecomm/MerchantHandler'